StakingContracts: refactoring for tests
